+++ apps/perms/utils/__init__.py
@@ -1,3 +1,4 @@
 from .permission import *
 from .user_permission import *
 from .account import *
+from .user_perm_tree import *
